LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-timeline-scroll-functional-notation.tentative-expected.txt


FAIL animation-timeline: scroll(nearest) assert_equals: expected "50px" but got "none"
FAIL animation-timeline: scroll(root) assert_equals: expected "50px" but got "none"
FAIL animation-timeline: scroll(self) assert_equals: expected "100px" but got "none"
PASS animation-timeline: scroll(self), on non-scroller
FAIL animation-timeline: scroll(inline) assert_equals: expected "100px" but got "none"
FAIL animation-timeline: scroll(x) assert_equals: expected "100px" but got "none"
FAIL animation-timeline: scroll(y) assert_equals: expected "100px" but got "none"
